+++ client/CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 ## 2021.07.14 [Unreleased]
 ### Added
 * Show visualization mode in the URL (map, table or chart) [LANDGRIF-119](https://vizzuality.atlassian.net/browse/LANDGRIF-119)
+* Show selected scenario in the URL [LANDGRIF-159](https://vizzuality.atlassian.net/browse/LANDGRIF-159)
+* Actual data as first element in the scenarios list [LANDGRIF-171](https://vizzuality.atlassian.net/browse/LANDGRIF-171)